+++ src/check_roms.rs
@@ -34,6 +34,14 @@ pub fn subcommand() -> Command {
                 .action(ArgAction::SetTrue),
         )
         .arg(
+            Arg::new("NAME")
+                .short('n')
+                .long("name")
+                .help("Select games by name")
+                .required(false)
+                .num_args(1),
+        )
+        .arg(
             Arg::new("SIZE")
                 .short('s')
                 .long("size")
@@ -49,6 +57,7 @@ pub async fn main(
     progress_bar: &ProgressBar,
 ) -> SimpleResult<()> {
     let systems = prompt_for_systems(connection, None, false, matches.get_flag("ALL")).await?;
+    let game_name = matches.get_one::<String>("NAME");
     let hash_algorithm = match find_setting_by_key(connection, "HASH_ALGORITHM")
         .await
         .unwrap()
@@ -66,6 +75,7 @@ pub async fn main(
             connection,
             progress_bar,
             &system,
+            &game_name,
             matches.get_flag("SIZE"),
             &hash_algorithm,
         )
@@ -79,11 +89,35 @@ async fn check_system(
     connection: &mut SqliteConnection,
     progress_bar: &ProgressBar,
     system: &System,
+    game_name: &Option<&String>,
     size: bool,
     hash_algorithm: &HashAlgorithm,
 ) -> SimpleResult<()> {
-    let header = find_header_by_system_id(connection, system.id).await;
-    let roms = find_roms_with_romfile_by_system_id(connection, system.id).await;
+    let games = match game_name {
+        Some(game_name) => {
+            let games = find_games_with_romfiles_by_name_and_system_id(
+                connection,
+                &format!("%{}%", game_name),
+                system.id,
+            )
+            .await;
+            prompt_for_games(games, cfg!(test))?
+        }
+        None => find_games_with_romfiles_by_system_id(connection, system.id).await,
+    };
+
+    if games.is_empty() {
+        if game_name.is_some() {
+            progress_bar.println(format!("No game matching \"{}\"", game_name.unwrap()));
+        }
+        return Ok(());
+    }
+
+    let roms = find_roms_with_romfile_by_game_ids(
+        connection,
+        &games.iter().map(|game| game.id).collect::<Vec<i64>>(),
+    )
+    .await;
     let romfiles = find_romfiles_by_system_id(connection, system.id).await;
     let mut roms_by_romfile_id: HashMap<i64, Vec<Rom>> = HashMap::new();
     roms.into_iter().for_each(|rom| {
@@ -92,6 +126,7 @@ async fn check_system(
             .or_default();
         group.push(rom);
     });
+    let header = find_header_by_system_id(connection, system.id).await;
 
     let mut transaction = begin_transaction(connection).await;