+++ src/check_roms/test_original_crc_mismatch.rs
@@ -60,6 +60,7 @@ async fn test() {
         &mut connection,
         &progress_bar,
         &system,
+        &None,
         false,
         &HashAlgorithm::Crc,
     )