+++ src/check_roms/test_rvz.rs
@@ -56,6 +56,7 @@ async fn test() {
         &mut connection,
         &progress_bar,
         &system,
+        &None,
         false,
         &HashAlgorithm::Crc,
     )