+++ src/check_roms/test_sevenzip_with_header.rs
@@ -58,6 +58,7 @@ async fn test() {
         &mut connection,
         &progress_bar,
         &system,
+        &None,
         false,
         &HashAlgorithm::Crc,
     )